requirements.txt
aspose-words==22.6.0
attrdict==2.0.1
babel==2.10.3
bce-python-sdk==0.8.73
cachetools==5.2.0
cfgv==3.3.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
cssselect==1.1.0
cssutils==2.4.2
cycler==0.11.0
cython==0.29.30
distlib==0.3.4
et-xmlfile==1.1.0
filelock==3.7.1
flake8==4.0.1
flask==2.1.2
flask-babel==2.0.0
fonttools==4.33.3
future==0.18.2
identify==2.5.1
idna==3.3
imageio==2.19.3
imgaug==0.4.0
importlib-metadata==4.11.4
itsdangerous==2.1.2
jinja2==3.1.2
kiwisolver==1.4.3
lmdb==1.3.0
lxml==4.9.0
markupsafe==2.1.1
matplotlib==3.5.2
mccabe==0.6.1
networkx==2.8.4
nodeenv==1.6.0
numpy==1.22.4
opencv-contrib-python==4.4.0.46
opencv-python==4.6.0.66
openpyxl==3.0.10
packaging==21.3
paddleocr==2.5.0.3
pandas==1.4.2
pillow==9.1.1
platformdirs==2.5.2
pre-commit==2.19.0
premailer==3.10.0
protobuf==4.21.1
pyclipper==1.3.0.post3
pycodestyle==2.8.0
pycryptodome==3.14.1
pyflakes==2.4.0
pygments==2.12.0
pyparsing==3.0.9
pytesseract==0.3.9
python-dateutil==2.8.2
python-levenshtein==0.12.2
pytz==2022.1
pywavelets==1.3.0
pyyaml==6.0
requests==2.28.0
rich==12.4.4
scikit-image==0.19.3
scipy==1.8.1
shapely==1.8.2
shellcheck-py==0.8.0.4
six==1.16.0
tifffile==2022.5.4
toml==0.10.2
tqdm==4.64.0
urllib3==1.26.9
virtualenv==20.14.1
visualdl==2.2.3
werkzeug==2.1.2
zipp==3.8.0
